Added Shake Multiplier
Added scripts that add a multiplier to the shake depending on how many blocks the user breaks within a certain time frame.
The script is a Singleton, and every time a block is broken it adds an update to the function. If the blocks break call is within the timer, it will add to the shake, which stacks with other calls. If it is outside of the timer, then it is a normal shake.
There are Serialized Fields in the GameManager to alter the base strength and duration of the shake, the multipliers of both, and the duration between block break calls